Inline util_path and util_chdir in util.c
